+++ Multiplatform/LiveKitExample.swift
@@ -23,117 +23,6 @@ import SwiftUI
                                               key: "preferences",
                                               default: Preferences())
 
-struct RoomSwitchView: View {
-    @EnvironmentObject var appCtx: AppContext
-    @EnvironmentObject var roomCtx: RoomContext
-    @EnvironmentObject var room: Room
-
-    #if os(visionOS)
-        @Environment(\.openImmersiveSpace) var openImmersiveSpace
-        @Environment(\.dismissImmersiveSpace) var dismissImmersiveSpace
-    #endif
-
-    var shouldShowRoomView: Bool {
-        room.connectionState == .connected || room.connectionState == .reconnecting
-    }
-
-    func computeTitle() -> String {
-        if shouldShowRoomView {
-            var elements: [String] = []
-            if let roomName = room.name {
-                elements.append(roomName)
-            }
-            if let localParticipantName = room.localParticipant.name {
-                elements.append(localParticipantName)
-            }
-            if let localParticipantIdentity = room.localParticipant.identity {
-                elements.append(String(describing: localParticipantIdentity))
-            }
-            return elements.joined(separator: " ")
-        }
-
-        return "LiveKit"
-    }
-
-    var body: some View {
-        ZStack {
-            Color.black
-                .ignoresSafeArea()
-
-            if shouldShowRoomView {
-                RoomView()
-            } else {
-                ConnectView()
-            }
-        }
-        .preferredColorScheme(.dark)
-        .navigationTitle(computeTitle())
-        .onChange(of: shouldShowRoomView) { newValue in
-            #if os(visionOS)
-                Task {
-                    if newValue {
-                        await openImmersiveSpace(id: "ImmersiveSpace")
-                    } else {
-                        await dismissImmersiveSpace()
-                    }
-                }
-            #endif
-        }
-    }
-}
-
-// Attaches RoomContext and Room to the environment
-struct RoomContextView: View {
-    @EnvironmentObject var appCtx: AppContext
-    @StateObject var roomCtx = RoomContext(store: sync)
-
-    var body: some View {
-        RoomSwitchView()
-            .environmentObject(roomCtx)
-            .environmentObject(roomCtx.room)
-            .foregroundColor(Color.white)
-            .onDisappear {
-                print("\(String(describing: type(of: self))) onDisappear")
-                Task {
-                    await roomCtx.disconnect()
-                }
-            }
-            .onOpenURL(perform: { url in
-
-                guard let urlComponent = URLComponents(url: url, resolvingAgainstBaseURL: false) else { return }
-                guard let host = url.host else { return }
-
-                let secureValue = urlComponent.queryItems?.first(where: { $0.name == "secure" })?.value?.lowercased()
-                let secure = ["true", "1"].contains { $0 == secureValue }
-
-                let tokenValue = urlComponent.queryItems?.first(where: { $0.name == "token" })?.value ?? ""
-
-                let e2ee = ["true", "1"].contains { $0 == secureValue }
-                let e2eeKey = urlComponent.queryItems?.first(where: { $0.name == "e2eeKey" })?.value ?? ""
-
-                var builder = URLComponents()
-                builder.scheme = secure ? "wss" : "ws"
-                builder.host = host
-                builder.port = url.port
-
-                guard let builtUrl = builder.url?.absoluteString else { return }
-
-                print("built URL: \(builtUrl), token: \(tokenValue)")
-
-                Task { @MainActor in
-                    roomCtx.url = builtUrl
-                    roomCtx.token = tokenValue
-                    roomCtx.isE2eeEnabled = e2ee
-                    roomCtx.e2eeKey = e2eeKey
-                    if !roomCtx.token.isEmpty {
-                        let room = try await roomCtx.connect()
-                        appCtx.connectionHistory.update(room: room, e2ee: e2ee, e2eeKey: e2eeKey)
-                    }
-                }
-            })
-    }
-}
-
 extension Decimal {
     mutating func round(_ scale: Int, _ roundingMode: NSDecimalNumber.RoundingMode) {
         var localCopy = self
